requirements.txt
transformers==4.5.1
numpy==1.18.5
pandas==1.2.4
xlrd==1.2.0
keras==2.3.1
tensorflow-gpu==2.3.0
h5py==2.10.0
scikit-learn==0.21.2